refactoring aggregation service
